Set up proper Rails engine tests, replace RSpec w. Minitest (#19)
Minitest tests are more familiar to Rails devs in general, and do not require as much mucking-about when testing a Rails engine. And we do want to test stepper_motor as a Rails engine because we need the database to be configured already when we run most of our tests - except for the tests of the generators / app setup.

Closes #16
Closes #17
Closes #4