Transactional outbox
There is a lot of back-and-forth going on in Rails core (or did go on there, from end of summer 2024 to December 2024) regarding whether the queue adapters should be able to define whether job enqueues are in-transaction (may co-commit) or will be shoved into after_commit.

This has severe implications for StepperMotor because it has no way of knowing whether the job that is going to trigger the step was actually enqueued or not. The main failure mode is the Rails process crashing between the COMMIT of the Journey changes (and eventually any other business logic stuff) and the after-commit hook that ActiveJob uses.

And since a Journey is "mutable" across multiple step invocations, we would not be able to know that it has been, indeed, scheduled.

So - while this bit of indirection is unnecessary if you have a co-commit DB ActiveJob queue, it _is_ necessary in most other cases. And it seems Rails core even goes as far as withholding the fact that you can (and should!) use co-commit if you can. There are also Sidekiqs and SQSes where co-commit is not possible at all.