Fix unique Journeys on MySQL (#35)
On MySQL/mariadb use a GENERATED column to stand in for the partial index used for uniqueness. It's not as slick and terse, but it does the job.

Add a migration which does the substitution of the index and only apply it with mysql adapters.

Add CI tests for Postgres and MySQL respectively.

Closes #14 
Closes #34