Add Journey-wide cancel_if (#41)
It is already the case in Cheddar that most journeys have a "blanket
stop" condition. Usually it pertains to the hero (the user, most of the
time) exiting the targeted cohort or no longer satisfying a certain
condition. To handle this, we add a blanket `cancel_if {}` at the class
level of the Journey, similar to Rails' `etag` method. The `cancel_if`
conditions will be evaluated before starting every step (hence the name:
"blanket" conditions).

```ruby
class UserOnboardingJourney < StepperMotor::Journey
  cancel_if { hero.deactivated? }
  cancel_if { hero.account_closed? }

  step :send_welcome_email do
    WelcomeMailer.welcome(hero).deliver_later
  end

  step :send_premium_offer, wait: 2.days do
    PremiumOfferMailer.exclusive_offer(hero).deliver_later
  end
end
```